Starter telemetry code
Some starter code for the telemetry parsing. The code using webrequest to read data from the server, parses it, stores it as an object, and shows how to access some fo the variables. Still not sure how we are planning on transmitting the suit/switch data from our server - hopefully this code can make the process easier.